feat: enhance RAG system with HNSW algorithm support and CLI utilities
Major improvements to the RAG (Retrieval-Augmented Generation) system:

Database & Storage Enhancements:
- Add HNSW (Hierarchical Navigable Small World) algorithm support for efficient vector similarity search
- Implement SQLiteVectorStoreHNSW with optimized HNSW graph structure
- Add new schema definitions for HNSW vector storage with Float32Array support
- Enhance database migration and configuration for HNSW parameters (M, Ml, EfSearch)

CLI Command Utilities:
- Add 'index' command for indexing plugins with configurable embedding models
- Add 'search' command for semantic search across indexed plugins
- Add 'list-collection' command to display all available vector collections
- Support for various embedding models and API configurations

API & Export Improvements:
- Simplify RAG API with cleaner, more intuitive interfaces
- Optimize plugin RAG manager with better metadata handling
- Enhance export/import functionality for vector collections
- Remove deprecated plugin_indexer.go in favor of new command structure

Algorithm & Performance:
- Implement pure Go HNSW algorithm with cosine similarity calculations
- Add distance function abstraction for different similarity metrics
- Optimize vector operations and graph traversal performance
- Add comprehensive test coverage for HNSW operations

Breaking Changes:
- Updated vector store schema to support HNSW-specific fields
- Modified collection configuration structure
- Enhanced embedding interface for better model compatibility

This update significantly improves the RAG system's search performance and provides
a complete CLI toolkit for managing vector collections and plugin indexing.

+++ common/ai/rag/plugins_rag/export_import_test.go
@@ -33,7 +33,7 @@ func TestMUSTPASS_ExportVectorData(t *testing.T) {
 	err = store.Add(rag.Document{
 		ID:      "Yakit 权威使用指南 v1",
 		Content: "Yakit 权威使用指南",
-		Embedding: []float64{
+		Embedding: []float32{
 			0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0,
 		},
 	})
@@ -43,7 +43,7 @@ func TestMUSTPASS_ExportVectorData(t *testing.T) {
 	err = store.Add(rag.Document{
 		ID:      "Yakit 权威使用指南 v2",
 		Content: "Yakit 权威使用指南",
-		Embedding: []float64{
+		Embedding: []float32{
 			0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0,
 		},
 	})
@@ -116,7 +116,7 @@ func TestMUSTPASS_ImportVectorData_ExistingCollection(t *testing.T) {
 	err = store.Add(rag.Document{
 		ID:      "original_doc",
 		Content: "Original document content",
-		Embedding: []float64{
+		Embedding: []float32{
 			0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0,
 		},
 	})
@@ -175,7 +175,7 @@ func TestMUSTPASS_ImportVectorData_ExistingDocument(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	originalEmbedding := []float64{0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0}
+	originalEmbedding := []float32{0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0}
 	err = store.Add(rag.Document{
 		ID:        "shared_doc_id",
 		Content:   "Original content",
@@ -209,7 +209,7 @@ func TestMUSTPASS_ImportVectorData_ExistingDocument(t *testing.T) {
 	err = store.Add(rag.Document{
 		ID:        "different_doc_id",
 		Content:   "Different document content",
-		Embedding: []float64{1.1, 1.2, 1.3, 1.4, 1.5, 1.6, 1.7, 1.8, 1.9, 2.0},
+		Embedding: []float32{1.1, 1.2, 1.3, 1.4, 1.5, 1.6, 1.7, 1.8, 1.9, 2.0},
 	})
 	assert.NoError(t, err)
 
@@ -262,7 +262,7 @@ func TestMUSTPASS_ImportVectorDataFullUpdate(t *testing.T) {
 	err = store.Add(rag.Document{
 		ID:      "Yakit 权威使用指南 v1",
 		Content: "Yakit 权威使用指南",
-		Embedding: []float64{
+		Embedding: []float32{
 			0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0,
 		},
 	})
@@ -280,7 +280,7 @@ func TestMUSTPASS_ImportVectorDataFullUpdate(t *testing.T) {
 	err = store.Add(rag.Document{
 		ID:      "Yakit 权威使用指南 v2",
 		Content: "Yakit 权威使用指南",
-		Embedding: []float64{
+		Embedding: []float32{
 			0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0,
 		},
 	})
